FIX: use int, not size_t for command line argument
git-svn-id: https://www.nitrc.org/svn/cmtk/trunk@1124 42a5c34f-2066-0410-bec5-ba365beb4995

+++ core/apps/reformatx.cxx
@@ -101,7 +101,7 @@ cmtk::XformList ReferenceToFloating;
 
 cmtk::ReformatVolume::Mode Mode = cmtk::ReformatVolume::REFORMAT_PLAIN;
 cmtk::Interpolators::InterpolationEnum Interpolation = cmtk::Interpolators::LINEAR;
-size_t InterpolatorWindowRadius = 3;
+int InterpolatorWindowRadius = 3;
 
 // default to auto-selection of data type
 cmtk::ScalarDataType DataType = cmtk::TYPE_NONE;
